+++ src/extensions/AssetView.ts
@@ -81,7 +81,7 @@ export class AssetView
     rootItem: AssetViewItem;
     path: string[] = [];
     rootPath: string = '';
-    selectedItem: AssetViewItem | undefined = undefined;
+    selectedItems: AssetViewItem[] = [];
     allowedTypes: any;
     searchValue: string = '';
     filter: string = 'None';
@@ -94,7 +94,8 @@ export class AssetView
     skipPreview: boolean = false;
     useNativeTitle: boolean = false;
     onlyFolders: boolean = true;
-    allowMultipleSelection: boolean = false;
+    allowMultipleSelection: boolean = true;
+    allowItemCheck: boolean = false;
     previewActions: any[] = [];
     contextMenu: any[] = [];
     itemContextMenuOptions: any = null;
@@ -148,6 +149,7 @@ export class AssetView
         this.useNativeTitle = options.useNativeTitle ?? this.useNativeTitle;
         this.onlyFolders = options.onlyFolders ?? this.onlyFolders;
         this.allowMultipleSelection = options.allowMultipleSelection ?? this.allowMultipleSelection;
+        this.allowItemCheck = options.allowItemCheck ?? this.allowItemCheck;
         this.previewActions = options.previewActions ?? [];
         this.itemContextMenuOptions = options.itemContextMenuOptions;
         this.gridScale = options.gridScale ?? this.gridScale;
@@ -322,7 +324,7 @@ export class AssetView
             itemEl.title = type + ': ' + item.id;
         }
 
-        if ( this.allowMultipleSelection )
+        if ( this.allowItemCheck )
         {
             let checkbox = document.createElement( 'input' );
             checkbox.type = 'checkbox';
@@ -440,13 +442,14 @@ export class AssetView
 
             if ( !isDoubleClick )
             {
-                if ( !e.shiftKey )
+                if ( !e.shiftKey || !that.allowMultipleSelection )
                 {
                     that.content.querySelectorAll( '.lexassetitem' ).forEach( ( i ) => i.classList.remove( 'selected' ) );
+                    that.selectedItems.length = 0;
                 }
 
                 this.classList.add( 'selected' );
-                that.selectedItem = item;
+                that.selectedItems.push( item );
 
                 if ( !that.skipPreview )
                 {
@@ -489,11 +492,11 @@ export class AssetView
             e.stopImmediatePropagation();
             e.stopPropagation();
 
-            const multiple = that.content.querySelectorAll( '.selected' ).length;
+            const multipleSelection = that.selectedItems.length > 1;
 
             const options: any[] = [
                 {
-                    name: ( multiple > 1 ) ? ( multiple + ' selected' ) : item.id,
+                    name: multipleSelection ? ( `${that.selectedItems.length} selected` ) : item.id,
                     icon: LX.makeIcon( 'CircleSmall', { svgClass: `fill-current text-${typeColor}` } ),
                     className: 'text-sm',
                     disabled: true
@@ -501,35 +504,42 @@ export class AssetView
                 null
             ];
 
-            if ( multiple <= 1 )
+            // By now, allow with none/single selected items
+            if ( !multipleSelection )
             {
                 options.push( { name: 'Rename', icon: 'TextCursor', callback: that._renameItemPopover.bind( that, item ) } );
             }
 
-            if ( !isFolder )
+            // By now, allow with none/single selected items
+            if ( !isFolder && !multipleSelection )
             {
                 options.push( { name: 'Clone', icon: 'Copy', callback: that._requestCloneItem.bind( that, item ) } );
             }
 
-            options.push( { name: 'Move', icon: 'FolderInput', callback: () => that._moveItem( item ) } );
+            // By now, allow with none/single selected items
+            if ( !multipleSelection )
+            {
+                options.push( { name: 'Move', icon: 'FolderInput', callback: () => that._moveItem( item ) } );
+            }
 
-            if ( type == 'Script' && LX.has( 'CodeEditor' ) )
+            // By now, allow with none/single selected items
+            if ( !multipleSelection && type == 'Script' && LX.has( 'CodeEditor' ) )
             {
                 options.push( { name: 'Open in Editor', icon: 'Code', callback: that._openScriptInEditor.bind( that, item ) } );
             }
 
             if ( that.itemContextMenuOptions )
             {
-                options.push( null );
+                if( options.length > 2 ) options.push( null );
 
                 for ( let o of that.itemContextMenuOptions )
                 {
                     if ( !o.name || !o.callback ) continue;
-                    options.push( { name: o.name, icon: o.icon, callback: o.callback?.bind( that, item ) } );
+                    options.push( { name: o.name, icon: o.icon, callback: o.callback?.bind( that, that.selectedItems.length ? that.selectedItems : [ item ] ) } );
                 }
             }
 
-            options.push( null, { name: 'Delete', icon: 'Trash2', className: 'destructive', callback: that._requestDeleteItem.bind( that, item ) } );
+            options.push( null, { name: 'Delete', icon: 'Trash2', className: 'destructive', callback: that._requestDeleteItem.bind( that, that.selectedItems.length ? that.selectedItems : [ item ] ) } );
 
             LX.addClass( that.contentPanel.root, 'pointer-events-none' );
 
@@ -811,7 +821,7 @@ export class AssetView
                     dom?.classList.add( 'selected' );
                 }
 
-                this.selectedItem = node;
+                this.selectedItems = [ node ];
             }
         } );
 
@@ -1007,6 +1017,7 @@ export class AssetView
         this.content.addEventListener( 'click', function()
         {
             this.querySelectorAll( '.lexassetitem' ).forEach( ( i ) => i.classList.remove( 'selected' ) );
+            that.selectedItems.length = 0;
         } );
 
         this.content.addEventListener( 'contextmenu', function( e )
@@ -1371,16 +1382,16 @@ export class AssetView
         return true;
     }
 
-    _requestDeleteItem( item: AssetViewItem )
+    _requestDeleteItem( items: AssetViewItem[] )
     {
         const onBeforeDelete = this._callbacks['beforeDelete'];
         const onDelete = this._callbacks['delete'];
 
         const resolve = ( ...args: any[] ) => {
-            this._deleteItem( item );
+            items.forEach( ( item ) => this._deleteItem( item ) );
             const event: AssetViewEvent = {
                 type: 'delete',
-                items: [ item ],
+                items,
                 userInitiated: true
             };
             if ( onDelete ) onDelete( event, ...args );
@@ -1390,7 +1401,7 @@ export class AssetView
         {
             const event: AssetViewEvent = {
                 type: 'delete',
-                items: [ item ],
+                items,
                 userInitiated: true
             };
             onBeforeDelete( event, resolve );